+++ installation-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### DEV
 
+ * Warm up the Symfony cache after the database credentials have been set (see #63).
  * Check if the Contao framework has been initialized when adding the user agent string (see standard-edition#64).
 
 ### 4.4.1 (2017-07-12)