+++ installation-bundle/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Contao installation bundle change log
 
-### DEV
+### 4.4.3 (2017-08-16)
 
  * Warm up the Symfony cache after the database credentials have been set (see #63).
  * Check if the Contao framework has been initialized when adding the user agent string (see standard-edition#64).